Move diff editor commands into their own files
